childControler class refactored
